implementing dynamic request

+++ TUM_IDP_Companion/TUM_IDP_Companion/Classes/Backend/BAL/GreenNavUrlRequest.h
@@ -20,6 +20,6 @@
 - (NSURLRequest*)urlRequestForNearestVerticeWithLatitude:(long)latitude longituide:(long)longituide;
 - (NSURLRequest*)urlRequestForVehicles;
 - (NSURLRequest*)urlRequestForVehiclesType:(NSString *)vehicleType;
-- (NSURLRequest*)urlRequestForVehicleRoutes:(NSString *)vehicle toRoute:(long)toRoute forRoute:(long)forRoute optimization:(NSString *)optimization battery:(NSUInteger)battery;
-- (NSURLRequest*)urlRequestForVehicleRoutes:(NSString *)vehicle range:(long)range battery:(NSUInteger)battery;
+- (NSURLRequest*)urlRequestForVehicleRoutes:(NSString *)vehicle toRoute:(long long)toRoute forRoute:(long long)forRoute optimization:(NSString *)optimization battery:(NSUInteger)battery;
+- (NSURLRequest*)urlRequestForVehicleRoutes:(NSString *)vehicle range:(long long)range battery:(NSUInteger)battery;
 @end
